% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel_eff.R
\name{fuel_eff}
\alias{fuel_eff}
\title{this function calculates fuel efficiency of a vehicle}
\usage{
fuel_eff(V, A, m, g = 9.8, p_air = 1.2, c_drag = 0.3, crolling = 0.015)
}
\arguments{
\item{V}{Velocity of the vehicle in m/s}

\item{A}{Surface area of the vehicle}

\item{m}{Mass of the vehicle}

\item{g}{Acceleration due to gravity}

\item{p_air}{Air Resistance}

\item{c_drag}{Drag from the vehicle}

\item{crolling}{I don't know what this is}
}
\description{
this function calculates fuel efficiency of a vehicle
}
